#53 🚚rename: ActivityHistory 및 AnimationWrapper 폴더 정리
