+++ frontend/src/components/common/AnimationWrapper.tsx
